staging v0.43.0

+++ transformslib/tables/collections/supply_load.py
@@ -1,6 +1,6 @@
 import os
 from typing import Dict, Any
-from transformslib.engine import get_engine, get_spark
+from transformslib.engine import get_engine, get_spark, detect_if_dbutils_available
 from tabulate import tabulate
 from adaptiveio import load_json
 from multitable import MultiTable, SchemaValidator
@@ -38,18 +38,20 @@ def get_execution_engine_info() -> Dict[str, Any]:
     ALL_VARS = os.environ
 
     #append with dbutils notebook info
-    try:
-        path = dbutils.notebook.entry_point.getDbutils().notebook().getContext().notebookPath().get()
-        ALL_VARS["DATABRICKS_NOTEBOOK_PATH"] = path
-        ALL_VARS["DATABRICKS_NOTEBOOK_NAME"] = os.path.basename(path)
-        ALL_VARS["DATABRICKS_WORKSPACE_URL"] = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiUrl().get()
-        ALL_VARS["DATABRICKS_CLUSTER_ID"] = dbutils.notebook.entry_point.getDbutils().notebook().getContext().tags().get("clusterId").get()
-        ALL_VARS["DATABRICKS_USER"] = dbutils.notebook.entry_point.getDbutils().notebook().getContext().tags().get("user").get()
-        
-    except NameError:
-        pass
-    except Exception as e:
-        print(f"SL020 Warning: Unhandled exception to get databricks notebook path: {e}")
+    dbav = detect_if_dbutils_available()
+    if dbav:
+        try:
+            path = dbutils.notebook.entry_point.getDbutils().notebook().getContext().notebookPath().get()
+            ALL_VARS["DATABRICKS_NOTEBOOK_PATH"] = path
+            ALL_VARS["DATABRICKS_NOTEBOOK_NAME"] = os.path.basename(path)
+            ALL_VARS["DATABRICKS_WORKSPACE_URL"] = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiUrl().get()
+            ALL_VARS["DATABRICKS_CLUSTER_ID"] = dbutils.notebook.entry_point.getDbutils().notebook().getContext().tags().get("clusterId").get()
+            ALL_VARS["DATABRICKS_USER"] = dbutils.notebook.entry_point.getDbutils().notebook().getContext().tags().get("user").get()
+            
+        except NameError:
+            pass
+        except Exception as e:
+            print(f"SL020 Warning: Unhandled exception to get databricks notebook path: {e}")
 
     return ALL_VARS
 
@@ -61,14 +63,12 @@ def clear_last_run():
         #remove it using 2 different systems
         path = transform_log_loc()
         if "dbfs:/" in path:
-            try:
-                #test if dbutils is available
-                dbls = dbutils.fs.ls("/")
-
+            dbav = detect_if_dbutils_available()
+            if dbav:
                 dbutils.fs.rm(path, True)
                 print("Cleared the transform log path using dbutils")
-            except NameError:
-                print("SL012 dbutils is NOT available to clear the path")
+            else:
+                raise NameError("SL012 dbutils could not clear the path")
         else:
             os.remove(path)
 
@@ -421,7 +421,11 @@ class SupplyLoad(TableCollection):
                 #explode the warnings on pipe
                 warnings_frame.explode("warning_messages", sep="|", outer=False)
                 warnings_frame = warnings_frame.get_pandas_frame()
-                warnings_frame = warnings_frame[warnings_frame["warning_messages"].notnull()]
+                # Filter out NULL AND empty strings
+                warnings_frame = warnings_frame[
+                    (warnings_frame["warning_messages"].notnull()) & 
+                    (warnings_frame["warning_messages"] != "")
+                ]
                 warnings_frame = warnings_frame.drop_duplicates()
                 print(tabulate(warnings_frame, headers='keys', tablefmt='pretty', showindex=False))
             except Exception as e:
